chore: fix dependencies in various tests
