+++ tests/test_skills/test_abstract_round_abci/test_common.py
@@ -42,9 +42,6 @@ from packages.valory.skills.abstract_round_abci.common import random_selection
 from packages.valory.skills.abstract_round_abci.test_tools.base import (
     FSMBehaviourBaseCase,
 )
-from packages.valory.skills.transaction_settlement_abci.payload_tools import (
-    VerificationStatus,
-)
 
 from tests.conftest import ROOT_DIR
 
@@ -324,7 +321,6 @@ class BaseRandomnessBehaviourTest(CommonBaseCase):
         assert self.behaviour.context.randomness_api._retries_attempted == 0
 
 
-# TOFIX: depends on transaction_settlement_abci, so should not be here in this form
 class BaseSelectKeeperBehaviourTest(CommonBaseCase):
     """Test SelectKeeperBehaviour."""
 
@@ -360,7 +356,6 @@ class BaseSelectKeeperBehaviourTest(CommonBaseCase):
                         dict(
                             participants=participants,
                             most_voted_randomness="56cbde9e9bbcbdcaf92f183c678eaa5288581f06b1c9c7f884ce911776727688",
-                            final_verification_status=VerificationStatus.PENDING,
                             blacklisted_keepers="".join(blacklisted_keepers),
                         )
                     ),
