+++ tests/test_skills/test_liquidity_rebalancing_abci/test_rounds.py
@@ -28,6 +28,7 @@ from packages.valory.skills.abstract_round_abci.base import AbciAppDB
 from packages.valory.skills.liquidity_rebalancing_abci.payloads import (
     SleepPayload,
     StrategyEvaluationPayload,
+    TransactionHashPayload,
 )
 from packages.valory.skills.liquidity_rebalancing_abci.rounds import (  # noqa: F401
     Event,
@@ -36,16 +37,11 @@ from packages.valory.skills.liquidity_rebalancing_abci.rounds import (  # noqa: 
     SynchronizedData,
     TransactionHashBaseRound,
 )
-
-# TOFIX - dep not allowed
-from packages.valory.skills.price_estimation_abci.payloads import TransactionHashPayload
 from packages.valory.skills.transaction_settlement_abci.payloads import ValidatePayload
 
 from tests.test_skills.test_abstract_round_abci.test_base_rounds import (
     BaseCollectSameUntilThresholdRoundTest,
 )
-
-# TOFIX - investigate if dep allowed
 from tests.test_skills.test_transaction_settlement_abci.test_rounds import (
     get_participant_to_signature,
 )
